requirements.txt
streamlit
openai
Pillow
matplotlib
